Refactor server mocks
This change will make it easier to write tests for the heartbeat
feature.
