+++ lib/dart_amqp.dart
@@ -4,4 +4,4 @@ export "src/client.dart";
 export "src/enums.dart";
 export "src/exceptions.dart";
 export "src/authentication.dart";
-export "src/protocol.dart" show MessageProperties;
+export "src/protocol.dart" show MessageProperties, TuningSettings;
